`printParenthesizedValueGroup`
