+++ src/language-css/printer-postcss.js
@@ -1,5 +1,4 @@
 import printString from "../utils/print-string.js";
-import isNextLineEmpty from "../utils/is-next-line-empty.js";
 import isNonEmptyArray from "../utils/is-non-empty-array.js";
 import {
   join,
@@ -7,13 +6,12 @@ import {
   hardline,
   softline,
   group,
-  fill,
   indent,
   dedent,
   ifBreak,
   breakParent,
 } from "../document/builders.js";
-import { removeLines, getDocParts } from "../document/utils.js";
+import { removeLines } from "../document/utils.js";
 import UnexpectedNodeError from "../utils/unexpected-node-error.js";
 import clean from "./clean.js";
 import embed from "./embed.js";
@@ -29,30 +27,24 @@ import {
   isLastNode,
   isSCSSControlDirectiveNode,
   isDetachedRulesetDeclarationNode,
-  isURLFunctionNode,
   hasComposesNode,
   hasParensAroundNode,
-  isKeyValuePairNode,
-  isKeyInValuePairNode,
   isDetachedRulesetCallNode,
   isTemplatePlaceholderNode,
   isTemplatePropNode,
-  isSCSSMapItemNode,
   isMediaAndSupportsKeywords,
   lastLineHasInlineComment,
-  isConfigurationNode,
-  isVarFunctionNode,
 } from "./utils/index.js";
 import { locStart, locEnd } from "./loc.js";
 import {
   adjustStrings,
   adjustNumbers,
   quoteAttributeValue,
-  shouldPrintComma,
   printUnit,
   printCssNumber,
-} from "./print/mic.js";
+} from "./print/misc.js";
 import printCommaSeparatedValueGroup from "./print/comma-separated-value-group.js";
+import printParenthesizedValueGroup from "./print/parenthesized-value-group.js";
 import printSequence from "./print/sequence.js";
 
 function genericPrint(path, options, print) {
@@ -498,123 +490,9 @@ function genericPrint(path, options, print) {
     case "value-comma_group":
       return printCommaSeparatedValueGroup(path, options, print);
 
-    case "value-paren_group": {
-      const parentNode = path.parent;
-      const printedGroups = path.map(() => {
-        const child = path.node;
-        return typeof child === "string" ? child : print();
-      }, "groups");
+    case "value-paren_group":
+      return printParenthesizedValueGroup(path, options, print);
 
-      if (
-        parentNode &&
-        isURLFunctionNode(parentNode) &&
-        (node.groups.length === 1 ||
-          (node.groups.length > 0 &&
-            node.groups[0].type === "value-comma_group" &&
-            node.groups[0].groups.length > 0 &&
-            node.groups[0].groups[0].type === "value-word" &&
-            node.groups[0].groups[0].value.startsWith("data:")))
-      ) {
-        return [
-          node.open ? print("open") : "",
-          join(",", printedGroups),
-          node.close ? print("close") : "",
-        ];
-      }
-
-      if (!node.open) {
-        return group(indent(fill(join([",", line], printedGroups))));
-      }
-
-      const isSCSSMapItem = isSCSSMapItemNode(path, options);
-
-      const lastItem = node.groups.at(-1);
-      const isLastItemComment = lastItem?.type === "value-comment";
-      const isKey = isKeyInValuePairNode(node, parentNode);
-      const isConfiguration = isConfigurationNode(node, parentNode);
-      const isVarFunction = isVarFunctionNode(parentNode);
-
-      const shouldBreak = isConfiguration || (isSCSSMapItem && !isKey);
-      const shouldDedent = isConfiguration || isKey;
-
-      const printed = group(
-        [
-          node.open ? print("open") : "",
-          indent([
-            softline,
-            join(
-              [line],
-              path.map(({ node: child, isLast, index }) => {
-                const hasComma = () =>
-                  Boolean(
-                    child.source &&
-                      options.originalText
-                        .slice(locStart(child), locStart(node.close))
-                        .trimEnd()
-                        .endsWith(",")
-                  );
-                const shouldPrintComma =
-                  !isLast || (isVarFunction && hasComma());
-                let printed = [
-                  printedGroups[index],
-                  shouldPrintComma ? "," : "",
-                ];
-
-                // Key/Value pair in open paren already indented
-                if (
-                  isKeyValuePairNode(child) &&
-                  child.type === "value-comma_group" &&
-                  child.groups &&
-                  child.groups[0].type !== "value-paren_group" &&
-                  child.groups[2]?.type === "value-paren_group"
-                ) {
-                  const parts = getDocParts(printed[0].contents.contents);
-                  parts[1] = group(parts[1]);
-                  printed = [group(dedent(printed))];
-                }
-
-                if (
-                  !isLast &&
-                  child.type === "value-comma_group" &&
-                  isNonEmptyArray(child.groups)
-                ) {
-                  let last = child.groups.at(-1);
-
-                  // `value-paren_group` does not have location info, but its closing parenthesis does.
-                  if (!last.source && last.close) {
-                    last = last.close;
-                  }
-
-                  if (
-                    last.source &&
-                    isNextLineEmpty(options.originalText, locEnd(last))
-                  ) {
-                    printed.push(hardline);
-                  }
-                }
-
-                return printed;
-              }, "groups")
-            ),
-          ]),
-          ifBreak(
-            !isLastItemComment &&
-              options.parser === "scss" &&
-              isSCSSMapItem &&
-              shouldPrintComma(options)
-              ? ","
-              : ""
-          ),
-          softline,
-          node.close ? print("close") : "",
-        ],
-        {
-          shouldBreak,
-        }
-      );
-
-      return shouldDedent ? dedent(printed) : printed;
-    }
     case "value-func":
       return [
         node.value,
